+++ src/pci/Cargo.toml
@@ -12,17 +12,12 @@ bench = false
 default = []
 
 [dependencies]
-byteorder = "1.5.0"
+
 displaydoc = "0.2.5"
 libc = "0.2.176"
 log = "0.4.28"
 serde = { version = "1.0.228", features = ["derive"] }
 thiserror = "2.0.17"
-vm-device = { path = "../vm-device" }
-vm-memory = { version = "0.16.1", features = [
-  "backend-mmap",
-  "backend-bitmap",
-] }
 
 [dev-dependencies]
 serde_test = "1.0.177"
